requirements.txt
antlr4-python3-runtime==4.9.3
audioread==3.0.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cloudpickle==3.1.1
colorama==0.4.6
decorator==5.2.1
demucs==4.0.1
dora_search==0.1.12
einops==0.8.1
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.13.1
fsspec==2024.6.1
future==1.0.0
idna==3.10
Jinja2==3.1.4
joblib==1.5.0
julius==0.2.7
lameenc==1.8.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
-e git+https://github.com/Alejandro-Dopico/Lyraoke-IA.git@a3d408fcb61ce432b27625e9f62c2462aa4d4331#egg=lyraoke
MarkupSafe==2.1.5
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.3
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu11==2.19.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
openai-whisper==20231117
openunmix==1.3.0
packaging==25.0
pillow==11.0.0
platformdirs==4.3.8
pooch==1.8.2
pycparser==2.22
pydub==0.25.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
scikit-learn==1.6.1
scipy==1.15.3
setuptools==70.2.0
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
submitit==1.5.2
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.9.0
torch==2.2.2+cu118
torchaudio==2.2.2+cu118
torchvision==0.20.1+cu121
tqdm==4.67.1
treetable==0.2.5
triton==2.3.1
typing_extensions==4.12.2
urllib3==2.4.0
Wave==0.0.2
